Modifications to kernel and jailhouse build to accomodate nouveau, and open source stack: initial attempts
See also: https://github.com/NVIDIA/tegra-nouveau-rootfs
Nvidia have a series of scripts and downstream repos that build
a complete open-source graphics stack (see link). Attempts have been
made here to refashion the scripts to be compatible with these scripts
but this has proved problematic.

Notes:
-Current downstream version of nouveau is only compatible with v.recent
kernel builds (>=4.7, 4.8 used here)
-Tegra-drm is an *integral* part. For the open-source stack
we need to display to the tegra-drm and render using nouveau.
- We can build nouveau 'in-tree' by copying the drm code directly
into the linux kernel
-Jailhouse builds are not seemingly compatible with using an output
directory for the linux build. Future commits will probably have to
revert this build process. Originally we did this for compatibility
with the scripts provided by Nvidia, although if we can build nouveau
directly, this is not needed as much.