Ability for systems to execute Jailhouse demonstrations on the TK1
Executing the UART demo is an important demonstration that non-root
cells are functioning correctly. This commit adds the inmate folder
to the jailhouse usr directory, as well as providing a script for
execution of the UART demo.

Also note that ttyS0 is configured so that console access is disabled.
This is required to prevent a hypervisor panic when concurrent access
between non-root and root cells occur.
https://blog.ramses-pyramidenbau.de/?p=342