Additional memory regions for TK1 root cell for HDMI function
With the existing memory map, hypervisor panics would occur upon
with an attached HDMI monitor. This revised memory map (from R. Ramsauer
https://groups.google.com/forum/#!msg/jailhouse-dev/iHXR2BLWI9M/Y8-eGy6XGwAJ
and http://pastebin.com/TEh5VXS5) solves this.

Note that HDMI does not function if the device is detached and reattached,
it has been tested with the device (HDMI monitor) already present. This
indicates that the plug-and-play capability is still somewhat lacking, but
it is a start

Also applied minor fix copying the freertos non-root cell descriptor over.