Adding dependencies and linkages required for openstack building
The nouveau stack consists of the drm driver (to be added in a later
commit), libdrm and mesa. Kmscube is a utility then used to test
this stack.

Mesa in particular has a large set of dependencies that need to be met.
These scripts (to be sourced) are to meet these dependencies